Change variables to 16 bit

+++ src/boot/ksdk1.1.0/pedometer.h
@@ -1,8 +1,8 @@
 #include <stdlib.h>
 
 typedef struct distribution{
-	int32_t mean;
-	int32_t variance;
+	int16_t mean;
+	int16_t variance;
 } distribution;
 
 typedef struct acc_distribution{
@@ -17,13 +17,13 @@ typedef struct acc_measurement{
 	int16_t z;
 } acc_measurement;
 
-int32_t compute_mean(int16_t data[], uint8_t N);
-int32_t compute_variance(int16_t data[], int32_t mean, uint8_t N);
+int16_t compute_mean(int16_t data[], uint8_t N);
+int16_t compute_variance(int16_t data[], int16_t mean, uint8_t N);
 distribution generate_distribution(int16_t data[], uint8_t N);
 acc_measurement read_accelerometer();
 acc_distribution read_acceleration_distribution(uint8_t N);
-void rotate_array_by_one(int32_t data[], uint8_t N);
-void print_array(int32_t data[], uint8_t N);
-void low_pass_filter(int32_t means[], int32_t vars[], uint8_t N, int32_t * output, int32_t * uncertainty);
+void rotate_array_by_one(int16_t data[], uint8_t N);
+void print_array(int16_t data[], uint8_t N);
+void low_pass_filter(int16_t means[], int16_t vars[], uint8_t N, int16_t * output, int16_t * uncertainty);
 void print_acc_distribution(acc_distribution dist);
 int8_t pedometer();
